Pass locale environment to initdb and pg_ctl
We now pass the environment dict containing locale settings to calls of
'initdb' and 'pg_ctl'. This is so as to avoid initdb to pick the locale
setting of its execution environment [1] which, when not an English one,
will then make the checks in PostgreSQLExecutor.running() (where the
message "pg_ctl: server is running" is expected) fail.

Fix #315.

[1]: https://www.postgresql.org/docs/current/locale.html